% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_diff.R
\name{binary_diff}
\alias{binary_diff}
\title{Binary Difference in Proportions}
\usage{
binary_diff(dt, num_col, reference, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{num_col}{the number of categorical columns in the data.}

\item{reference}{the name of the reference row category to use.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with difference in proportions test results between pairs of columns for binary data, as well as an overall chi-squared test across all groups.
}
\description{
Default comparison function for binary data.
}
\keyword{tangram.pipe}
