% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_jama.R
\name{binary_jama}
\alias{binary_jama}
\title{JAMA-style summary for a Binary Row}
\usage{
binary_jama(dt, reference, rowlabel, compact, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{reference}{the name of the row category to use as the reference. Default will use alphabetical first category.}

\item{rowlabel}{the label for the table row name, if different from row_var.}

\item{compact}{logical: if TRUE, data displayed in one row.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with summary statistics for a binary variable.
}
\description{
Summarizes a binary row using column percentages and the total number in each cell divided by the column total. This is the stye used by the Journal of the American Medical Association.
}
\keyword{tangram.pipe}
