% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_rr.R
\name{binary_rr}
\alias{binary_rr}
\title{Binary Risk Ratio}
\usage{
binary_rr(dt, num_col, reference, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{num_col}{the number of categorical columns in the data.}

\item{reference}{the name of the reference row category to use.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with computed risk ratios between pairs of columns for binary data, as well as an overall chi-squared test across all groups.
}
\description{
Calculates risk ratio across categories for binary data.
}
\keyword{tangram.pipe}
