% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_comp_default.R
\name{cat_comp_default}
\alias{cat_comp_default}
\title{Chi-Squared Test for Categorical Variables}
\usage{
cat_comp_default(dt, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe calculating relative entropy between column pairs, as well as an overall chi-squared test across all groups.
}
\description{
Default comparison function for categorical data.
}
