% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_entry.r
\name{cat_entry}
\alias{cat_entry}
\title{Add a categorical variable summary row to the table}
\usage{
cat_entry(tbl, x, y = initial_approach, dt = data, xlab = NULL,
  fun = NULL, ...)
}
\arguments{
\item{tbl}{The table object.  (This input is generally not specified because the tbl object is passed via the \code{\link[magrittr]{\%>\%}} operator.  See examples.)}

\item{x}{Name of the row categorical variable.  This is the non-standard implementation, so do not quote the name or provide it with a character string object.}

\item{y}{Name of the column categorical variable.  Do not provide a character string.  Generally, this parameter is specified in the table header and not in this function.}

\item{dt}{The data.table where the variables x and y reside.  Generally, this parameter is specified in the table header and not in this function.}

\item{xlab}{Character string.  Text will be the label of the categorical entry in the table.  If not specified, function defaults to \code{label(x)} if available.  As a last resort, the function will print the variable name.}

\item{fun}{Function which summarizes the \code{table(x, y, usaNA = "always")} object.  Defaults to \code{count_pct} which provides \code{N (pct)}.}

\item{...}{Options passed to fun.  These generally include pvalue and fmt.}
}
\description{
This function adds a categorical variable summary row to a table object.
}
\keyword{table}
