% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_jama.R
\name{cat_jama}
\alias{cat_jama}
\title{JAMA-style summary for a Categorical Row}
\usage{
cat_jama(dt, rowlabel, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{rowlabel}{the label for the table row name, if different from row_var.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with summary statistics for a categorical variable.
}
\description{
Summarizes a categorical row using column percentages and the total number in each cell divided by the column total. This is the stye used by the Journal of the American Medical Association.
}
\keyword{tangram.pipe}
