% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_pct.R
\name{cat_pct}
\alias{cat_pct}
\title{Percentage summary for a Categorical Row}
\usage{
cat_pct(dt, rowlabel, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{rowlabel}{the label for the table row name, if different from row_var.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with summary statistics for a categorical variable.
}
\description{
Summarizes a categorical row using counts and column percentages.
}
\keyword{tangram.pipe}
