% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_row.R
\name{empty_row}
\alias{empty_row}
\title{Empty Row}
\usage{
empty_row(list_obj, header = NULL)
}
\arguments{
\item{list_obj}{the name of the tbl_start object previously initialized.}

\item{header}{a header to include for the empty row.}
}
\description{
Produces a empty dividing row in the table.  May have a row header.
}
\keyword{tangram.pipe}
