% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{indent}
\alias{indent}
\title{Indentation}
\usage{
indent(df, spaces)
}
\arguments{
\item{df}{The output data frame object to be printed in HTML form}

\item{spaces}{The number of spaces for the indentation}
}
\description{
Used to put indentation for categorical row categories for HTML-style table
}
