% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris2-data.R
\docType{data}
\name{iris2}
\alias{iris2}
\title{Tangram.pipe Iris Data}
\format{
\code{iris2} is a data frame with 500 rows sampled from the original \code{iris}
dataset. In addition, three made-up categorical variables are introduced to
the dataset, named \code{Stem.Size}, \code{Leaf.Color}, and \code{color}.
}
\source{
The \code{iris} dataset.
}
\usage{
data(iris2)
}
\description{
This dataset is adapted from Edgar Anderson's Iris Data for use in the
tangram.pipe package. Three new variables (stem size, leaf color, and flower color)
are introduced into the dataset as made-up categorical variables which were not
from Fisher's original data. The dataset is expanded to 500 by sampling rows
from the original 150 row \code{iris} dataset.
}
\keyword{datasets}
