% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_row.R
\name{n_row}
\alias{n_row}
\title{Row counter}
\usage{
n_row(
  list_obj,
  col_var = NULL,
  newdata = FALSE,
  missing = NULL,
  overall = NULL
)
}
\arguments{
\item{list_obj}{the name of the tbl_start object previously initialized.}

\item{col_var}{the variable to be used in the table columns. Default is from initialized tbl_start object.}

\item{newdata}{enter new dataset name if different from that initialized in tbl_start.}

\item{missing}{logical: if TRUE, missing data in the column variable is considered; FALSE only uses complete cases.}

\item{overall}{logical: if TRUE, an overall column is included.}
}
\value{
A list with the row counts added as a new element to \code{list_obj}.
}
\description{
Counts the instances of each column variable of the dataframe to be used in
the table (if applicable), and gives an overall row count.
}
\examples{
x <- tbl_start(iris, "Species", missing=TRUE, overall=TRUE, comparison=TRUE) \%>\%
  n_row()
}
\keyword{tangram.pipe}
