% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_medianiqr.R
\name{num_medianiqr}
\alias{num_medianiqr}
\title{Median/IQR summary for a Numeric Row}
\usage{
num_medianiqr(dt, rowlabel, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{rowlabel}{the label for the table row name, if different from row_var.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with summary statistics for a numeric variable.
}
\description{
Summarizes a numeric row using the median and interquartile range.
}
\keyword{tangram.pipe}
