% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_minmax.R
\name{num_minmax}
\alias{num_minmax}
\title{Min-Max summary for a Numeric Row}
\usage{
num_minmax(dt, rowlabel, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{rowlabel}{the label for the table row name, if different from row_var.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\value{
A dataframe with summary statistics for a numeric variable.
}
\description{
Summarizes a numeric row using the minimum and maximum values.
}
\keyword{tangram.pipe}
