% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_row.R
\name{num_row}
\alias{num_row}
\title{Numeric Row}
\usage{
num_row(
  list_obj,
  row_var,
  col_var = NULL,
  newdata = FALSE,
  rowlabel = NULL,
  summary = NULL,
  missing = NULL,
  overall = NULL,
  comparison = NULL,
  digits = 2
)
}
\arguments{
\item{list_obj}{the name of the \code{tbl_start} object previously initialized.}

\item{row_var}{the name of the variable to be used in the rows.}

\item{col_var}{the variable to be used in the table columns. Default is from initialized \code{tbl_start} object.}

\item{newdata}{enter new dataset name if different from that initialized in \code{tbl_start}.}

\item{rowlabel}{the label for the table row name, if different from \code{row_var}.}

\item{summary}{summary function for the data, if different from the one supplied in \code{tbl_start}.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{overall}{logical: if TRUE, an overall column is included.}

\item{comparison}{the name of the comparison test to use, if different from that initialized in \code{tbl_start}.}

\item{digits}{significant digits to use.}
}
\value{
A list with the numeric row's table information added as a new element to \code{list_obj}.
}
\description{
Adds in a numeric row to a \code{tangram.pipe} table.
}
\examples{
x <- tbl_start(iris, "Species", missing=TRUE, overall=TRUE, comparison=TRUE) \%>\%
  num_row("Sepal.Length", rowlabel="Sepal Length")
}
\seealso{
Possible summary functions for numeric data: \link[tangram.pipe]{num_default}, \link[tangram.pipe]{num_mean_sd}, \link[tangram.pipe]{num_medianiqr}, \link[tangram.pipe]{num_minmax}, \link[tangram.pipe]{num_date}

Other related row-building functions: \link[tangram.pipe]{cat_row}, \link[tangram.pipe]{binary_row}, \link[tangram.pipe]{n_row}, \link[tangram.pipe]{empty_row}

Starting a \code{tangram.pipe} table: \link[tangram.pipe]{tbl_start}
}
\keyword{tangram.pipe}
