% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tangram.pipe.R
\name{print.tangram.pipe}
\alias{print.tangram.pipe}
\title{Printing a Table}
\usage{
\method{print}{tangram.pipe}(list_obj)
}
\arguments{
\item{list_obj}{the name of the tbl_start object previously initialized.}
}
\description{
Prints a finished table
}
\keyword{tangram.pipe}
