% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangram_styling.R
\name{tangram_styling}
\alias{tangram_styling}
\title{Tangram Styling}
\usage{
tangram_styling(df)
}
\arguments{
\item{df}{The output data frame object to be printed in HTML form.}
}
\value{
A dataframe containing HTML formatting code where applicable.
}
\description{
Used to preprocess a tangram.pipe table for HTML formatting.
}
