% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_print.R
\name{tbl_print}
\alias{tbl_print}
\title{Printing a Table}
\usage{
tbl_print(list_obj)
}
\arguments{
\item{list_obj}{the name of the tbl_start object previously initialized.}
}
\description{
Prints a finished table
}
\keyword{tangram.pipe}
