% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_start.R
\name{tbl_start}
\alias{tbl_start}
\title{Table Initialization}
\usage{
tbl_start(data, col_var, missing, overall, comparison)
}
\arguments{
\item{data}{the dataset to be used in the table.}

\item{col_var}{the variable to be used in the table columns. NULL if single summary column desired}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{overall}{logical: if TRUE, an overall column is included.}

\item{comparison}{logical: if TRUE, a comparison test is conducted between columns.}
}
\description{
Initializes the table by specifying the desired elements and data components.
}
\examples{
x <- tbl_start(iris, "Species", missing=TRUE, overall=TRUE, comparison=TRUE)
}
\keyword{tangram.pipe}
