% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_jama.R
\name{binary_jama}
\alias{binary_jama}
\title{JAMA-style summary for a Binary Row}
\usage{
binary_jama(dt, ...)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{...}{Additional arguments supplied within the package row functions.}
}
\value{
A dataframe with summary statistics for a binary variable.
}
\description{
Summarizes a binary row using column percentages and the total number in each cell divided by the column total. This is the style used by the Journal of the American Medical Association.
}
\details{
This is an internal function of \code{tangram.pipe}. Additional arguments
should be supplied for this function to work properly.

\code{reference} : the name of the row category to use as the reference. Default will use alphabetical first category

\code{rowlabel} : the label for the table row name, if different from row_var.

\code{compact} : if TRUE, data displayed in one row.

\code{missing} : if TRUE, missing data is considered; FALSE only uses complete cases.

\code{digits} : significant digits to use.
}
\seealso{
Possible summary functions for binary data:\link[tangram.pipe]{binary_default}, \link[tangram.pipe]{binary_pct}, \link[tangram.pipe]{binary_count}
}
\keyword{tangram.pipe}
