% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pct.R
\name{binary_pct}
\alias{binary_pct}
\title{Percentage summary for a Binary Row}
\usage{
binary_pct(dt, ...)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{...}{Additional arguments supplied within the package row functions.}
}
\value{
A dataframe with summary statistics for a binary variable.
}
\description{
Summarizes a binary row using counts and column percentages.
}
\details{
This is an internal function of \code{tangram.pipe}. Additional arguments
should be supplied for this function to work properly.

\code{reference} : the name of the row category to use as the reference. Default will use alphabetical first category

\code{ref.label} : choice of whether you want the reference label to be in the table. Default is \code{on} and includes reference label; \code{off} switches it off.

\code{rowlabel} : the label for the table row name, if different from row_var.

\code{compact} : if TRUE, data displayed in one row.

\code{missing} : if TRUE, missing data is considered; FALSE only uses complete cases.

\code{digits} : significant digits to use.
}
\seealso{
Possible summary functions for binary data:\link[tangram.pipe]{binary_default}, \link[tangram.pipe]{binary_count}, \link[tangram.pipe]{binary_jama}
}
\keyword{tangram.pipe}
