% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_default.R
\name{cat_default}
\alias{cat_default}
\title{Summary for a Categorical Row}
\usage{
cat_default(dt, row_var, rowlabels, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{row_var}{the name of the variable to be used in the rows.}

\item{rowlabels}{the label for the table row name, if different from row_var.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\description{
Summarizes a categorical row using counts and column proportions.
}
\keyword{tangram.pipe}
