% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_default.R
\name{cat_default}
\alias{cat_default}
\title{Default summary for a Categorical Row}
\usage{
cat_default(dt, ...)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{...}{Additional arguments supplied within the package row functions.}
}
\value{
A dataframe with summary statistics for a categorical variable.
}
\description{
Summarizes a categorical row using counts and column proportions.
}
\details{
This is an internal function of \code{tangram.pipe}. Additional arguments
should be supplied for this function to work properly.

\code{rowlabel} : the label for the table row name, if different from row_var.

\code{missing} : if TRUE, missing data is considered; FALSE only uses complete cases.

\code{digits} : significant digits to use.
}
\keyword{tangram.pipe}
