% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_default.R
\name{num_default}
\alias{num_default}
\title{Summary for a Numeric Row}
\usage{
num_default(dt, rowlabels, missing, digits)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{rowlabels}{the label for the table row name, if different from row_var.}

\item{missing}{logical: if TRUE, missing data is considered; FALSE only uses complete cases.}

\item{digits}{significant digits to use.}
}
\description{
Summarizes a numeric row using the five-number summary, mean, and standard deviation.
}
\keyword{tangram.pipe}
