% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_default.R
\name{num_default}
\alias{num_default}
\title{Default summary for a Numeric Row}
\usage{
num_default(dt, ...)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{...}{Additional arguments supplied within the package row functions.}
}
\value{
A dataframe with summary statistics for a numeric variable.
}
\description{
Summarizes a numeric row using the five-number summary, mean, and standard deviation.
}
\details{
This is an internal function of \code{tangram.pipe}. Additional arguments
should be supplied for this function to work properly.

\code{rowlabel} : the label for the table row name, if different from row_var.

\code{missing} : if TRUE, missing data is considered; FALSE only uses complete cases.

\code{digits} : significant digits to use.
}
\seealso{
Additional prewritten summary functions for numeric data: \link[tangram.pipe]{num_mean_sd}, \link[tangram.pipe]{num_medianiqr}, \link[tangram.pipe]{num_minmax}, \link[tangram.pipe]{num_date}
}
\keyword{tangram.pipe}
