% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_mean_sd.R
\name{num_mean_sd}
\alias{num_mean_sd}
\title{Mean/SD summary for a Numeric Row}
\usage{
num_mean_sd(dt, ...)
}
\arguments{
\item{dt}{the name of the dataframe object.}

\item{...}{Additional arguments supplied within the package row functions.}
}
\value{
A dataframe with summary statistics for a numeric variable.
}
\description{
Summarizes a numeric row using the mean and standard deviation.
}
\details{
This is an internal function of \code{tangram.pipe}. Additional arguments
should be supplied for this function to work properly.

\code{rowlabel} : the label for the table row name, if different from row_var.

\code{missing} : if TRUE, missing data is considered; FALSE only uses complete cases.

\code{digits} : significant digits to use.
}
\keyword{tangram.pipe}
