% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_out.R
\name{tbl_out}
\alias{tbl_out}
\title{Output Table}
\usage{
tbl_out(list_obj)
}
\arguments{
\item{list_obj}{the name of the tbl_start object previously initialized.}
}
\description{
Produces a finalized table
}
\keyword{tangram.pipe}
